Rename lib module to mapping
