+++ lib/mapping/cache.js
