+++ lib/mapping/doc-info-adapter.js
