+++ lib/mapping/index.js
