+++ lib/mapping/model-batch-item.js
