+++ lib/mapping/model-mapping-info.js
