+++ lib/mapping/object-selector.js
