+++ lib/mapping/query-generator.js
