+++ lib/mapping/result-mapper.js
