+++ lib/mapping/result.js
