+++ lib/mapping/table-mappings.js
