+++ lib/mapping/tree.js
